Rename folder containing FirebaseClient
