+++ pwa/src/FirebaseClient/FirebaseClient.js
